Epiphanies were had!
Changed the variable that hold am or pm to a boolean expression. (true for am, false for pm)
Fixed all the parameters and names that this change broke.
Added condition statement to timeTick() method that tests each time an hour is incremented. If the hour is equal to zero the boolean switches.