Fixed updateDisplay() to work for 12Hr Clock Display.
This is where the epiphany came in.
Added condition statements to updateDisplay() to test the value of the boolean meridian so it can print out am or pm correctly.
Added a nested condition statement to simply print out 12 if hour value is zero.
added comments to hopefully make it clearer.